+++ lib/Extension/CoreExtension.php
@@ -21,8 +21,8 @@ use PhpBench\Benchmark\Telespector;
 use PhpBench\Console\Application;
 use PhpBench\Console\Command\ReportCommand;
 use PhpBench\Console\Command\RunCommand;
-use PhpBench\Container;
-use PhpBench\ExtensionInterface;
+use PhpBench\DependencyInjection\Container;
+use PhpBench\DependencyInjection\ExtensionInterface;
 use PhpBench\ProgressLogger\DotsProgressLogger;
 use PhpBench\ProgressLogger\VerboseProgressLogger;
 use PhpBench\ProgressLoggerRegistry;
