+++ lib/Progress/Logger/DotsLogger.php
