+++ lib/Progress/Logger/NullProgress.php
