+++ lib/Progress/LoggerInterface.php
