+++ lib/Progress/LoggerRegistry.php
