+++ lib/Report/GeneratorInterface.php
